Improve rendering speed
Instead of re-drawing the entire grid on each frame, we split it into
a 4*4 grid. When a tile type changes, we mark rendering grid cell that
it is contained in as stale, and re-draw its texture. If no cells within
a rendering grid cell change, then the last texture is copied as-is.